+++ src/neuroconv/tools/nwb_helpers/__init__.py
@@ -1,0 +1,9 @@
+from ._metadata_and_file_helpers import (
+    add_device_from_metadata,
+    get_module,
+    get_default_nwbfile_metadata,
+    make_nwbfile_from_metadata,
+    make_or_load_nwbfile,
+)
+from ._dataset_and_backend_models import ConfigurableDataset, DatasetConfiguration, BackendConfiguration
+from ._dataset_configuration import get_configurable_datasets, get_default_dataset_configurations